Convert site to Hugo
As part of #98, we decided that we wanted to replace the single HTML
file with a static site generator.

This allows for easier adding/changing of content and theme, as
currently the setup of the site will make it a little difficult to
change the structure of the site in the future, and it also makes it
easier for folks to contribute changes to the events on the site.

We've decided to use Hugo, as a lightweight solution that only requires
a single download to get it working.

To make this work and automagically deploy on merges, we need to use
GitHub Actions.

We no longer need anything Gulp/Node related, as Hugo can minify our
source for us, so we can remove those references.

Closes #98.